Island: Move FlaskDIWrapper into monkey_island.cc.flask_utils
