+++ monkey/monkey_island/cc/flask_utils/abstract_resource.py
