+++ monkey/monkey_island/cc/resources/auth/authenticate.py
@@ -5,7 +5,7 @@ import flask_jwt_extended
 from flask import make_response, request
 
 from common.utils.exceptions import IncorrectCredentialsError
-from monkey_island.cc.resources.AbstractResource import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.resources.auth.credential_utils import get_username_password_from_request
 from monkey_island.cc.resources.request_authentication import create_access_token
 from monkey_island.cc.services import AuthenticationService
