+++ monkey/monkey_island/cc/resources/auth/registration_status.py
@@ -1,4 +1,4 @@
-from monkey_island.cc.resources.AbstractResource import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.services import AuthenticationService
