+++ monkey/monkey_island/cc/resources/clear_simulation_data.py
@@ -3,7 +3,7 @@ from http import HTTPStatus
 from flask import make_response
 
 from monkey_island.cc.event_queue import IIslandEventQueue, IslandEventTopic
-from monkey_island.cc.resources.AbstractResource import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.resources.request_authentication import jwt_required
