+++ monkey/monkey_island/cc/resources/exploitations/monkey_exploitation.py
@@ -1,11 +1,11 @@
 from dataclasses import asdict
 
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.repositories import (
     IAgentEventRepository,
     IAgentPluginRepository,
     IMachineRepository,
 )
-from monkey_island.cc.resources.AbstractResource import AbstractResource
 from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.reporting.exploitations.monkey_exploitation import (
     get_monkey_exploited,
