+++ monkey/monkey_island/cc/resources/local_run.py
@@ -2,7 +2,7 @@ import json
 
 from flask import jsonify, make_response, request
 
-from monkey_island.cc.resources.AbstractResource import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.run_local_monkey import LocalMonkeyRunService
