+++ monkey/monkey_island/cc/resources/machines.py
@@ -1,7 +1,7 @@
 from http import HTTPStatus
 
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.repositories import IMachineRepository
-from monkey_island.cc.resources.AbstractResource import AbstractResource
 from monkey_island.cc.resources.request_authentication import jwt_required
