+++ monkey/monkey_island/cc/resources/ransomware_report.py
@@ -1,11 +1,11 @@
 from flask import jsonify
 
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.repositories import (
     IAgentEventRepository,
     IAgentPluginRepository,
     IMachineRepository,
 )
-from monkey_island.cc.resources.AbstractResource import AbstractResource
 from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.ransomware import ransomware_report
