+++ monkey/monkey_island/cc/services/agent_configuration_service/flask_resources/agent_configuration_schema.py
@@ -1,6 +1,6 @@
 from http import HTTPStatus
 
-from monkey_island.cc.resources.AbstractResource import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource
 from monkey_island.cc.services import IAgentConfigurationService
